+++ compose.yml
@@ -13,7 +13,6 @@ x-service-template: &service-template
     - node_modules_ghost_core:/home/ghost/ghost/core/node_modules:delegated
     - node_modules_ghost_custom-fonts:/home/ghost/ghost/custom-fonts/node_modules:delegated
     - node_modules_ghost_domain-events:/home/ghost/ghost/domain-events/node_modules:delegated
-    - node_modules_ghost_donations:/home/ghost/ghost/donations/node_modules:delegated
     - node_modules_ghost_email-addresses:/home/ghost/ghost/email-addresses/node_modules:delegated
     - node_modules_ghost_email-service:/home/ghost/ghost/email-service/node_modules:delegated
     - node_modules_ghost_html-to-plaintext:/home/ghost/ghost/html-to-plaintext/node_modules:delegated
@@ -178,7 +177,6 @@ volumes:
   node_modules_ghost_core: {}
   node_modules_ghost_custom-fonts: {}
   node_modules_ghost_domain-events: {}
-  node_modules_ghost_donations: {}
   node_modules_ghost_email-addresses: {}
   node_modules_ghost_email-service: {}
   node_modules_ghost_html-to-plaintext: {}
