+++ ghost/core/core/server/services/stripe/services/webhook/CheckoutSessionEventService.js
@@ -1,4 +1,3 @@
-const {DonationPaymentEvent} = require('@tryghost/donations');
 const _ = require('lodash');
 const errors = require('@tryghost/errors');
 const logging = require('@tryghost/logging');
@@ -63,6 +62,7 @@ module.exports = class CheckoutSessionEventService {
         const memberRepository = this.deps.memberRepository;
         const member = session.customer ? (await memberRepository.get({customer_id: session.customer})) : null;
 
+        const {DonationPaymentEvent} = require('../../../donations/DonationPaymentEvent');
         const data = DonationPaymentEvent.create({
             name: member?.get('name') ?? session.customer_details.name,
             email: member?.get('email') ?? session.customer_details.email,
