+++ ghost/core/package.json
@@ -78,7 +78,6 @@
     "@tryghost/database-info": "0.3.29",
     "@tryghost/debug": "0.1.34",
     "@tryghost/domain-events": "0.0.0",
-    "@tryghost/donations": "0.0.0",
     "@tryghost/email-addresses": "0.0.0",
     "@tryghost/email-mock-receiver": "0.3.10",
     "@tryghost/email-service": "0.0.0",
