--- ghost/donations/src/DonationRepository.ts
@@ -1,5 +1,0 @@
-import {DonationPaymentEvent} from './DonationPaymentEvent';
-
-export type DonationRepository = {
-    save(event: DonationPaymentEvent): Promise<void>;
-}
