--- ghost/donations/src/index.ts
@@ -1,3 +1,0 @@
-export * from './DonationPaymentEvent';
-export * from './DonationRepository';
-export * from './DonationBookshelfRepository';
