--- ghost/donations/test/hello.test.ts
@@ -1,8 +1,0 @@
-import assert from 'assert/strict';
-
-describe('Hello world', function () {
-    it('Runs a test', function () {
-        // TODO: Write me!
-        assert.ok(require('../'));
-    });
-});
